refactor: 메서드 추출
